Sensitivity analysis of diagnostic meta-analysis with fixed c

DTA sensitivity analysis with pre-specified (fixed) c1 and c2

dtametasa.fc(
  data,
  p,
  c1.sq = 0.5,
  correct.value = 0.5,
  correct.type = "all",
  brem.init = NULL,
  b.init = 1,
  b.interval = c(0, 2),
  a.interval = c(-3, 3),
  neg.r = FALSE,
  ci.level = 0.95,
  show.warn.message = FALSE,
  a.root.extendInt = "downX",
  ...
)

Arguments

data
Data with variable names either {TP, FN, TN, FP} or {y1, y2, v1, v2}. If not, please change the variable names. Either data with the number of subjects or logit transformed data works.

p
Specified probability of selection (or publication); Pr(select) = p

c1.sq
Pre-specified \(c_1^2\). Default value is 0.5; hence, \(c_2^2 = 1 - c_1^2\)

correct.value
Imputation value for ``continuity correction''. Default is 0.5; When zero cell exists, add some value.

correct.type
2 ways for ``continuity correction''. Default is all: add correct.value into all the cells of the row with zero value; if single: add correct.value into only the cells with zero value.

brem.init
Initial values used for estimating the parameters in the bivariate random effects model. Default is NA, which set the estimation without considering publication bias as initial values; It should be changed by a vector of u1 u2 t1 t2 r, c(u1, u2, t1, t2, r). Bad initial values will cause non-convergence results.

b.init
An initial value of \(\beta\). Default is 0.1. Avoid to start from 0. Bad initial value will cause non-convergence results.

b.interval
The constraint interval for \(\beta\). Default is in the interval c(0,2). Positive side (>0) is recommended to be adopted. The estimation of \(\beta\) will be searched within the interval.

a.interval
The constraint interval for \(\alpha\). Default is in the interval c(-3,3). then, the root of \(\alpha\) will be searched within the interval.

neg.r
Whether to permit correlation parameter \(\rho\) to be a negative value. Default is FALSE, \(\rho\) in [-1, 1]. If TRUE, \(\rho\) is in [-1, 0)

ci.level
The significant value for confidence interval. Default is 0.95, hence, a 2-tailed 95 profile likelihood.

show.warn.message
Whether to show the warning messages. Default is to hide warning messages.

a.root.extendInt
See extendInt augment in function uniroot.

...
See other augments in function uniroot.

Value

confidence interval, convergence list, logit transformed data

Details

Continuity correction

https://en.wikipedia.org/wiki/Continuity_correction

See also

Examples

sa1 <- dtametasa.fc(IVD, p = 0.7)
sa1

#> $par.all
#>         u1         u2        t11        t22        t12         c1         c2 
#>  1.2986358  1.7169013  0.3042539  0.7088328 -0.1591841  0.7071068  0.7071068 
#>          b          a       sauc         se         sp 
#>  2.0000000 -4.3077014  0.8361984  0.7856053  0.8477293 
#>