Sensitivity analysis of diagnostic meta-analysis with prespecified contrast vector

Sensitivity analysis with pre-specified (fixed) contrast vector \(\boldsymbol{c} = (c_1, c_2)\)

dtametasa.fc(
  data,
  p,
  c1.square = 0.5,
  beta0 = 1,
  reitsma.par0 = NULL,
  beta.interval = c(0, 2),
  alpha.interval = c(-3, 3),
  ci.level = 0.95,
  correct.value = 0.5,
  correct.type = c("single", "all")[1],
  show.warn.message = FALSE,
  alpha.root.extendInt = "downX",
  eps = sqrt(.Machine$double.eps),
  sauc.type = c("sroc", "hsroc")[1],
  ...
)

Arguments

data

Data with variable names either {TP, FN, TN, FP} or {y1, y2, v1, v2}. If not, please change the variable names. Either data with the number of subjects or logit transformed data works.

p

Specified probability of selection (or publication); Pr(select) = p

c1.square

Pre-specified \(c_1^2\). \(c_2^2 = 1 - c_1^2\)

beta0

An initial value of \(\beta\). Avoid to start from 0. Bad initial value will cause non-convergence results.

reitsma.par0

Initial values used for estimating the parameters in the bivariate random effects model (Reitsma's model). It should be changed by a vector of c(mu1, mu2, tau1, tau2, rho). Bad initial values will cause non-convergence results. Default is NULL, which uses estimated results from Reitsma's model without taking into account the publication bias.

beta.interval

The constraint interval for \(\beta\). The estimation of \(\beta\) will be searched within the interval. Take either the positive interval (>0) or the negative interval (<0).

alpha.interval

The constraint interval for \(\alpha\). then, the root of \(\alpha\) will be searched within the interval.

ci.level

The significant value for confidence interval of SAUC.s

correct.value

Imputation value for ``continuity correction''.

correct.type

Two types of ``continuity correction''. "single": input value for single study. "all": input value for all the cells.

show.warn.message

Whether to show the warning messages.

alpha.root.extendInt

See extendInt augment in function uniroot.

eps

A minimum positive value.

sauc.type

Two types of SAUC values. Area under the Reitsma's SROC curve("sroc") or under the Rutter's HSROC curve ("hsroc").

...

See other augments in function uniroot.

Value

confidence interval, convergence list, logit transformed data

Details

Continuity correction

https://en.wikipedia.org/wiki/Continuity_correction

See also

Examples

sa.fit1 <- dtametasa.fc(IVD, p = 0.7)
sa.fit1
#> $par.all
#>    mu1    mu2 tau1^2 tau2^2  tau12   c1^2   c2^2   beta  alpha   sroc   sens 
#>  1.380  1.740  0.355  0.713 -0.178  0.500  0.500  2.000 -4.290  0.851  0.799 
#>   spec 
#>  0.851 
#> 

sa.fit2 <- dtametasa.fc(IVD, p = 0.7, sauc.type = "sroc", correct.type = "all")
sa.fit2
#> $par.all
#>    mu1    mu2 tau1^2 tau2^2  tau12   c1^2   c2^2   beta  alpha   sroc   sens 
#>  1.299  1.717  0.304  0.709 -0.159  0.500  0.500  2.000 -4.308  0.836  0.786 
#>   spec 
#>  0.848 
#> 

sa.fit3 <- dtametasa.fc(IVD, p = 0.7, sauc.type = "hsroc")
sa.fit3
#> $par.all
#>    mu1    mu2 tau1^2 tau2^2  tau12   c1^2   c2^2   beta  alpha  hsroc   sens 
#>  1.380  1.740  0.355  0.713 -0.178  0.500  0.500  2.000 -4.290  0.889  0.799 
#>   spec 
#>  0.851 
#>